Properly deref flag to fix metaschema-deprecated-types per metaschema-framework/metaschema-java#406
Per metaschema-framework/metaschema-java#407 for metaschema-framework/metaschema-java#406,
we need to fix the constraint that is serialized into Java code. To
properly align the constraints and ensure parity, we need to backport
the fix for others not using the Java-based handwritten constraints or
want to run the same best practices constraints 'out of band' with the
CLI or another conformant runtime.